Reorganized directories for easier management
