+++ client/src/pages/about/assets/staff_picture.png
