+++ client/src/pages/authentification/VerifiedPage.js
