+++ client/src/pages/destinationpage/components/BookingInfo.js
