+++ client/src/pages/destinationpage/components/DescriptionComponent.js
